UGATIT/requirements.txt
boto3
matplotlib
ipython

absl-py==0.9.0
astor==0.8.0
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
certifi==2020.6.20
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
gast==0.2.2
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
importlib-metadata==1.6.1
ipykernel==5.3.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
mistune==0.8.4
#mkl-fft==1.1.0
#mkl-random==1.1.1
#mkl-service==2.3.0
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numpy==1.18.1
olefile==0.46
opt-einsum==3.1.0
packaging==20.4
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pyzmq==19.0.1
QtPy==1.9.0
scipy==1.4.1
Send2Trash==1.5.0
six==1.15.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
wcwidth==0.2.4
webencodings==0.5.1
Werkzeug==0.16.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0
opencv-python
